+++ src/modules/VammModule.sol
@@ -31,7 +31,7 @@ contract VammModule is IVammModule {
     /**
      * @inheritdoc IVammModule
      */
-    function configureVamm(uint128 _marketId, uint256 _maturityTimestamp, VammConfiguration.Mutable calldata _config)
+    function configureVamm(uint128 _marketId, uint32 _maturityTimestamp, VammConfiguration.Mutable calldata _config)
     external override
     {
         OwnableStorage.onlyOwner();
